Fix staging startup and harden OWASP NVD caching in CI.
Align JPA entity IDs with Flyway BIGINT columns, fix the seed BCrypt hash, and make Docker staging smoke pass; add resilient NVD cache workflows to survive NIST 524 timeouts.

+++ src/main/java/com/companyx/equity/error/TransactionNotFoundException.java
@@ -3,7 +3,7 @@ package com.companyx.equity.error;
 public class TransactionNotFoundException extends RuntimeException {
     private static final long serialVersionUID = 1L;
 
-    public TransactionNotFoundException(String uid, Integer transactionId) {
+    public TransactionNotFoundException(String uid, Long transactionId) {
         super("Transaction not found for user: " + uid + ", transactionId: " + transactionId);
     }
 }
