+++ src/main/java/com/companyx/equity/repository/TransactionRepository.java
@@ -13,29 +13,29 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface TransactionRepository extends JpaRepository<Transaction, Integer> {
+public interface TransactionRepository extends JpaRepository<Transaction, Long> {
     
     @Query("SELECT t FROM Transaction t WHERE t.user.id = :uid AND t.id = :id")
     @QueryHints(@QueryHint(name = "org.hibernate.cacheable", value = "false"))
-    Optional<Transaction> findByUidAndId(@Param("uid") int uid, @Param("id") int id);
+    Optional<Transaction> findByUidAndId(@Param("uid") Long uid, @Param("id") Long id);
 
     @Query("SELECT t FROM Transaction t WHERE t.user.id = :uid AND t.timestamp < :end ORDER BY t.timestamp")
     @QueryHints({
         @QueryHint(name = "org.hibernate.fetchSize", value = "100"),
         @QueryHint(name = "org.hibernate.readOnly", value = "true")
     })
-    List<Transaction> findAllBefore(@Param("uid") int uid, @Param("end") Date end);
+    List<Transaction> findAllBefore(@Param("uid") Long uid, @Param("end") Date end);
 
     @Query("SELECT t FROM Transaction t WHERE t.user.id = :uid AND t.timestamp BETWEEN :start AND :end ORDER BY t.timestamp")
     @QueryHints({
         @QueryHint(name = "org.hibernate.fetchSize", value = "100"),
         @QueryHint(name = "org.hibernate.readOnly", value = "true")
     })
-    List<Transaction> findAllBetween(@Param("uid") int uid, @Param("start") Date start, @Param("end") Date end);
+    List<Transaction> findAllBetween(@Param("uid") Long uid, @Param("start") Date start, @Param("end") Date end);
     
     @Query("SELECT COUNT(t) FROM Transaction t WHERE t.user.id = :uid")
-    long countByUserId(@Param("uid") int uid);
+    long countByUserId(@Param("uid") Long uid);
 
     @Query("SELECT MIN(t.timestamp) FROM Transaction t WHERE t.user.id = :uid AND t.symbol = :symbol")
-    Optional<Date> findEarliestByUserAndSymbol(@Param("uid") int uid, @Param("symbol") String symbol);
+    Optional<Date> findEarliestByUserAndSymbol(@Param("uid") Long uid, @Param("symbol") String symbol);
 }
