+++ src/main/java/com/companyx/equity/repository/TransactionTypeRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TransactionTypeRepository extends JpaRepository<TransactionType, Integer> {
+public interface TransactionTypeRepository extends JpaRepository<TransactionType, Long> {
 }
